[JBPAPP-9517]:Fix multiple nested schemas in wsdl can not validate message
git-svn-id: https://anonsvn.jboss.org/repos/jbossws/stack/native/branches/jbossws-native-3.1.2@16474 50344849-fb0f-0410-a53f-a94597aac05c

+++ modules/testsuite/native-tests/src/test/java/org/jboss/test/ws/jaxws/jbws1172/types/PerformTest.java
@@ -1,24 +1,4 @@
-/*
- * JBoss, Home of Professional Open Source.
- * Copyright 2006, Red Hat Middleware LLC, and individual contributors
- * as indicated by the @author tags. See the copyright.txt file in the
- * distribution for a full listing of individual contributors.
- *
- * This is free software; you can redistribute it and/or modify it
- * under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation; either version 2.1 of
- * the License, or (at your option) any later version.
- *
- * This software is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this software; if not, write to the Free
- * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
- * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
- */
+
 package org.jboss.test.ws.jaxws.jbws1172.types;
 
 import javax.xml.bind.annotation.XmlAccessType;
@@ -37,7 +17,7 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent>
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *       &lt;sequence>
- *         &lt;element name="Code" type="{http://www.w3.org/2001/XMLSchema}long"/>
+ *         &lt;element name="Code" type="{http://www.my-company.it/ws/my-test2}CodeType"/>
  *       &lt;/sequence>
  *     &lt;/restriction>
  *   &lt;/complexContent>
@@ -52,18 +32,18 @@ import javax.xml.bind.annotation.XmlType;
 })
 public class PerformTest {
 
-    @XmlElement(name = "Code", required = true, type = Long.class, nillable = true)
-    protected Long code;
+    @XmlElement(name = "Code", required = true, type = Integer.class, nillable = true)
+    protected Integer code;
 
     /**
      * Gets the value of the code property.
      * 
      * @return
      *     possible object is
-     *     {@link Long }
+     *     {@link Integer }
      *     
      */
-    public Long getCode() {
+    public Integer getCode() {
         return code;
     }
 
@@ -72,10 +52,10 @@ public class PerformTest {
      * 
      * @param value
      *     allowed object is
-     *     {@link Long }
+     *     {@link Integer }
      *     
      */
-    public void setCode(Long value) {
+    public void setCode(Integer value) {
         this.code = value;
     }
